Fix terminal response not sent during sim hotswap.
In a case, where single sim is removed and reinserted,
terminal response for the subsequent display text/select
item proactive command is not sent, which results in GCF
test case failure.

Fix: When sim card is absent initialize the CatService instance
to null, so that StkAppService always use the latest CatService
instance when sim card becomes present.

Change-Id: Id667d7497c55b46e256eada05aabe3a3b2b9a6b6
CRs-Fixed: 2172791